Set up Flutter folder structure and add dependencies
