+++ pubspec.yaml
@@ -1,0 +1,25 @@
+name: alltune
+description: A universal musical instrument tuning app built with Flutter.
+version: 0.1.0
+publish_to: 'none'
+
+environment:
+  sdk: '>=2.17.0 <3.0.0'
+
+dependencies:
+  flutter:
+    sdk: flutter
+  flutter_fft: ^2.4.0
+  flutter_animate: ^4.1.1
+  shared_preferences: ^2.2.2
+  google_mobile_ads: ^3.0.0
+
+dev_dependencies:
+  flutter_test:
+    sdk: flutter
+
+flutter:
+  uses-material-design: true
+  assets:
+    - assets/images/
+    - assets/audio/
